角色卡 / 夥伴卡整合 + 攝影機平移動畫 + 視覺層級修正
角色卡 / 夥伴卡：
- 新增 Skill 模型（Active/Passive）；Character 加 Description/Skills；
  NpcCompanion 加 Description/Personality/SceneResponses/Skills
- 刪除無用的 CharacterCard/CompanionCard α 包裝型別（資料統一走底層）
- EquipmentSlot.Character 標 Obsolete；CharacterCardSlot 預設改 Head
- EquipmentService.MoveCharacterCard 改寫成「與裝備卡相同」的 swap 規則
  + 新增 SwapCharacterWithEquipmentSlot / GetTotalStats
- WorldMap 暴露 Character/Companion + LoadCharacter/LoadCompanion +
  MoveCharacterCardToSlot；裝備路徑自動阻擋角色卡所在槽 / 背包
- 新增 CharacterCardView（金邊「角色」卡，只顯示四屬性 2×2）+ 擴充
  EquipmentSlotView 承載角色卡（drag kind=character_card，禁背包）
- LeftPanel 主角區 / 夥伴區改 data-driven，移除寫死值
- EventResolutionDialog 加夥伴提示 stub（依事件 scene 對 SceneResponses）
- 補 schema 新欄位 + abandoned-mansion / valid-module 範例資料
- MainBootstrap 載入 demo 角色（scholar）+ 夥伴（old-priest）

攝影機平移動畫：
- Projection.ProjectQuad 簽章從 int 改 float（body 不動，原本就用
  DepthSliceContinuous，已支援小數）
- MainMapRenderer 加 _cameraSubTileOffset（rel 座標小數補間，非螢幕像素）
  + AnimateCameraToPlayer：起始 (newRow-oldRow, newCol-oldCol) tween
  到 Vector2.Zero，0.25s Cubic EaseOut
- UpdateAllTiles / UpdateParallaxScene 改吃 float rel，每幀重新透視投影
  → 遠近列各自依 perspective 換算位置，平滑鏡頭 pan
- 視野判定加 0.5 容差避免邊緣 tile pop；Reset 鍵會 kill 進行中 tween

視覺層級修正（drag preview / popup 不被立繪盒擋）：
- ParallaxScene 為 ZIndex=5；以下層級全部明確 ZIndex 蓋過
- drag preview wrapper：4096（CardView / EquipmentSlotView 共用 helper）
- ActionTriggerPopup：20（行動選單）
- PlayCardZone：10（drop hint + 出牌動畫）
- 修 EquipmentSlotView.SetEquipment 角色卡引用沒清乾淨，導致角色卡
  移動時舊槽殘留複製品的 bug

測試：新增 8 個（角色卡 swap / 拒絕背包 / 4 屬性總和 / 新欄位反序列化），
全套 344 測試綠。

+++ core/Map/Projection.cs
@@ -58,10 +58,10 @@ public static class Projection
     }
 
     /// <summary>計算單一地塊以「地面 1×1 單位」投影出的 4 角梯形（後緣窄、前緣寬）。</summary>
-    /// <param name="relRow">玩家為原點的相對 row：-2 (最遠) ~ +2 (最近)</param>
-    /// <param name="relCol">玩家為原點的相對 col：-2 (左) ~ +2 (右)</param>
+    /// <param name="relRow">玩家為原點的相對 row：-2 (最遠) ~ +2 (最近)。可傳小數以支援攝影機補間。</param>
+    /// <param name="relCol">玩家為原點的相對 col：-2 (左) ~ +2 (右)。可傳小數以支援攝影機補間。</param>
     /// <param name="p">投影參數</param>
-    public static ProjectedQuad ProjectQuad(int relRow, int relCol, ProjectionParams p)
+    public static ProjectedQuad ProjectQuad(float relRow, float relCol, ProjectionParams p)
     {
         // 後緣（深度 = relRow - 0.5）與前緣（深度 = relRow + 0.5）
         var (yBack, scaleBack) = DepthSliceContinuous(relRow - 0.5f, p);
